Added support for stdincludes.txt and stddefines.txt
Though stddefines.txt is just a stub for now, because it requires
randomdude's pull request. Added both files to changelog and manual
(again just a stup for stddefines.txt). Added a test which makes use of
both features (currently fails because of missing stddefines.txt
functionality) and made test suite actually generate the required txt
files for that (that was the only reasonable solution I could come up
with, since the contents of stdincludes.txt are dependent on its
location and the location of the tests directory).