Fix a bug with the string replace function
Closes #111.

Issue #111 was caused by the string replace function not "doing its job"
when removing duplicated spaces from input. That was caused by the
dequote macro returning early due to an incomplete string (no closing
quote). That incomplete string was created because the null terminator
was erroneously added to the output string. (i'm not 100% sure about
that last part, but fixing that seems to fix the other bugs, so...)