update travis ci build environment
this should, first of all, update gcc from 5.4 to 7.4, but also update
wine to 3.0. This might fix the crash on 64-bit mingw, as I found a
bug report with a very similar error message and reproduction steps
which was apparently fixed in wine 1.7.40, whereas the old build
environment had wine 1.6.2.