Fix bitness dependency
strtoul gives one result on 64bit Unix, and another result on everything else (both 32bit Unix and 64bit Windows), for nine-digit inputs (for example dd $FFFFFFFF00000000>>16). strtoull will give the same result for everything.

strtoull is in C99 and C++11, so it should exist on every compiler we care about.