Set PDB_NAME to something other than asar
CMake was automatically setting the PDB_NAME to be asar.pdb, which on Windows conflicts with the dlls pdb name, since it's the same (asar.dll -> asar.pdb). This was giving arcane linker errors since it was trying to write the same file at the same time from 2 places and it completely broke debugging on Visual Studio, since it wasn't loading the debug symbols. Setting the PDB_NAME alone works fine, symbols load correctly and there are no linker errors. 
This was also tested on Linux to confirm that it works just fine, cmake just ignores the property but doesn't complain.