Build asar as a static library too.
With these changes, building asar now also builds a static library version of it.
The name of the library will be asar-static.lib/libasar-static.a .
This behavior is enabled by default (as everything else) and can be disabled.
To do this, add -DASAR_GEN_LIB=OFF to the cmake configuration command.