Normalize paths before inserting them in the virtual filesystem
If we don't do this, then memory files passed using patchparams to asar_patch_ex won't be found unless they are equal to their normalized form. 
E.g. if you pass a memory file with a path "Test.asm", and a patch has "incsrc Test.asm", asar will check if "test.asm" exists in the vfs and it won't find it, leading to an error.