The Freespace Rewrite: Part 2
allow making rats-tag-less freespaces (with the segment command). also
uhh. fix some jank with freedata static i think? or maybe just fix
something i broke in the last commit. also fix the freespace finder
trying to overlap freespace with code from the same patch (this didn't
turn up with the existing freespace finder only because nobody was crazy
enough to try both using freespace and writing to "the freespace region"
manually with org. but i think this is a reasonable thing to do with
segments so i made it work properly)

(as a side effect, even writing zeros will now cause the freespace
finder to not touch that part. this may or may not affect some use
cases, idk. one that i can think of is padding the rom size. maybe we
should add a separate command for that instead)

this has been the 2:30am git commit ramble, thanks for reading