allow label refs in skip,fill,pad,incbin
`org Label` is allowed already, and none of these can cause more
breakage than that, so that's not an excuse to ban them. and there are
apparently some quite valid use cases for them

(well, technically for pad the old behavior was to allow all labels,
including forward ones. there's no way to use that for anything useful
though, so i banned that too)